move queue to scheduler package
